+++ src/app/blog/3d_printer/page.mdx
@@ -3,7 +3,7 @@ import { Accordion, AccordionDetails, AccordionSummary, Container } from "@mui/m
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import DarkModeFix from '@/components/muiWrappers/darkModeFix/_darkModeFix';
 import Lazy from "@/components/lazy/_lazy";
-import { generateCustomMetadata } from "@/content/Metadata";
+import { generateCustomMetadata } from "@/Metadata";
 
 import logo from "./opengraph-image.png";
 import s1 from "./images/s1.png";
