+++ src/app/blog/city_fibre/page.mdx
@@ -3,7 +3,7 @@ import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import DarkModeFix from '@/components/muiWrappers/darkModeFix/_darkModeFix';
 import Lazy from "@/components/lazy/_lazy";
 import LazyLocalImage from "@/components/lazy/_lazyLocalImage";
-import { generateCustomMetadata } from "@/content/Metadata";
+import { generateCustomMetadata } from "@/Metadata";
 
 import logo from "./opengraph-image.png";
 import s0 from "./images/s0.png";
